pyproject.toml
[project]
name = "traitbank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.9.0",
    "attrs>=25.3.0",
    "ichatbio-sdk>=0.2.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
]
